Factor out common parts to base ccomparisionBusinesscase
